fixed 64-bit release build
Got rid of a ton of size_t usages in core library

+++ clr/DNP3CLRAdapter/Conversions.cpp
@@ -159,7 +159,7 @@ OctetString^ Conversions::convertMeas(const opendnp3::OctetString& arMeas)
 {
 	auto buffer = arMeas.ToReadOnly();
 	array<System::Byte>^ bytes = gcnew array<System::Byte>(buffer.Size());
-	for(size_t i = 0; i< buffer.Size(); ++i) bytes[i] = buffer[i];
+	for (uint32_t i = 0; i< buffer.Size(); ++i) bytes[i] = buffer[i];
 	return gcnew OctetString(bytes);	
 }
