+++ cpp/opendnp3/include/opendnp3/ICommandProcessor.h
@@ -43,7 +43,7 @@ public:
 	* @param aIndex index of the command
 	* @param aCallback callback that will be invoked upon completion or failure
 	*/
-	virtual void SelectAndOperate(const ControlRelayOutputBlock& arCommand, size_t aIndex, std::function<void (CommandResponse)> aCallback) = 0;
+	virtual void SelectAndOperate(const ControlRelayOutputBlock& arCommand, uint16_t aIndex, std::function<void (CommandResponse)> aCallback) = 0;
 
 	/**
 	* Direct operate a ControlRelayOutputBlock
@@ -51,7 +51,7 @@ public:
 	* @param aIndex index of the command
 	* @param aCallback callback that will be invoked upon completion or failure
 	*/
-	virtual void DirectOperate(const ControlRelayOutputBlock& arCommand, size_t aIndex, std::function<void (CommandResponse)> aCallback) = 0;
+	virtual void DirectOperate(const ControlRelayOutputBlock& arCommand, uint16_t aIndex, std::function<void (CommandResponse)> aCallback) = 0;
 
 	/**
 	* Select and operate a 16 bit analog output
@@ -59,7 +59,7 @@ public:
 	* @param aIndex index of the command
 	* @param aCallback callback that will be invoked upon completion or failure
 	*/
-	virtual void SelectAndOperate(const AnalogOutputInt16& arCommand, size_t aIndex, std::function<void (CommandResponse)> aCallback) = 0;
+	virtual void SelectAndOperate(const AnalogOutputInt16& arCommand, uint16_t aIndex, std::function<void (CommandResponse)> aCallback) = 0;
 
 	/**
 	* Direct operate a 16 bit analog output
@@ -67,7 +67,7 @@ public:
 	* @param aIndex index of the command
 	* @param aCallback callback that will be invoked upon completion or failure
 	*/
-	virtual void DirectOperate(const AnalogOutputInt16& arCommand, size_t aIndex, std::function<void (CommandResponse)> aCallback) = 0;
+	virtual void DirectOperate(const AnalogOutputInt16& arCommand, uint16_t aIndex, std::function<void (CommandResponse)> aCallback) = 0;
 
 	/**
 	* Select and operate a 32 bit analog output
@@ -75,7 +75,7 @@ public:
 	* @param aIndex index of the command
 	* @param aCallback callback that will be invoked upon completion or failure
 	*/
-	virtual void SelectAndOperate(const AnalogOutputInt32& arCommand, size_t aIndex, std::function<void (CommandResponse)> aCallback) = 0;
+	virtual void SelectAndOperate(const AnalogOutputInt32& arCommand, uint16_t aIndex, std::function<void (CommandResponse)> aCallback) = 0;
 
 	/**
 	* Direct operate a 32 bit analog output
@@ -83,7 +83,7 @@ public:
 	* @param aIndex index of the command
 	* @param aCallback callback that will be invoked upon completion or failure
 	*/
-	virtual void DirectOperate(const AnalogOutputInt32& arCommand, size_t aIndex, std::function<void (CommandResponse)> aCallback) = 0;
+	virtual void DirectOperate(const AnalogOutputInt32& arCommand, uint16_t aIndex, std::function<void (CommandResponse)> aCallback) = 0;
 
 	/**
 	* Select and operate a single precision analog output
@@ -91,7 +91,7 @@ public:
 	* @param aIndex index of the command
 	* @param aCallback callback that will be invoked upon completion or failure
 	*/
-	virtual void SelectAndOperate(const AnalogOutputFloat32& arCommand, size_t aIndex, std::function<void (CommandResponse)> aCallback) = 0;
+	virtual void SelectAndOperate(const AnalogOutputFloat32& arCommand, uint16_t aIndex, std::function<void (CommandResponse)> aCallback) = 0;
 
 	/**
 	* Direct operate a single precision analog output
@@ -99,7 +99,7 @@ public:
 	* @param aIndex index of the command
 	* @param aCallback callback that will be invoked upon completion or failure
 	*/
-	virtual void DirectOperate(const AnalogOutputFloat32& arCommand, size_t aIndex, std::function<void (CommandResponse)> aCallback) = 0;
+	virtual void DirectOperate(const AnalogOutputFloat32& arCommand, uint16_t aIndex, std::function<void (CommandResponse)> aCallback) = 0;
 
 	/**
 	* Select and operate a double precision analog output
@@ -107,7 +107,7 @@ public:
 	* @param aIndex index of the command
 	* @param aCallback callback that will be invoked upon completion or failure
 	*/
-	virtual void SelectAndOperate(const AnalogOutputDouble64& arCommand, size_t aIndex, std::function<void (CommandResponse)> aCallback) = 0;
+	virtual void SelectAndOperate(const AnalogOutputDouble64& arCommand, uint16_t aIndex, std::function<void (CommandResponse)> aCallback) = 0;
 
 	/**
 	* Direct operate a double precision analog output
@@ -115,7 +115,7 @@ public:
 	* @param aIndex index of the command
 	* @param aCallback callback that will be invoked upon completion or failure
 	*/
-	virtual void DirectOperate(const AnalogOutputDouble64& arCommand, size_t aIndex, std::function<void (CommandResponse)> aCallback) = 0;
+	virtual void DirectOperate(const AnalogOutputDouble64& arCommand, uint16_t aIndex, std::function<void (CommandResponse)> aCallback) = 0;
 };
 
 }
